cmd/codegen/generator/go: use astscan + schematool in generate-module
Fold Phase-1 (source analysis) and Phase-2 (schema merge) into
generate-module so Go modules no longer need a separate
generate-typedefs pass. The AST-based scanner replaces the old
typedef-extraction path; schema merging happens in-process via
the schematool library.

The legacy packages.Load-based typedef path is retained behind a
//go:build legacy_typedefs tag (generate_typedefs_legacy.go) for
parity comparison and will be dropped once parity is confirmed.

Adds --legacy-typedefs and --self-calls flags to the codegen CLI
to enable opt-in to the legacy path or the experimental SELF_CALLS
feature respectively.
