+++ cmd/codegen/generator/go/generate_typedefs.go
@@ -1,137 +1,21 @@
+//go:build !legacy_typedefs
+// +build !legacy_typedefs
+
 package gogenerator
 
 import (
 	"context"
 	"fmt"
-	"go/token"
-	"io/fs"
-	"os"
-	"os/exec"
-	"path/filepath"
 
 	"github.com/dagger/dagger/cmd/codegen/generator"
-	"github.com/dagger/dagger/cmd/codegen/generator/go/templates"
 	"github.com/dagger/dagger/cmd/codegen/introspection"
-	"github.com/dschmidt/go-layerfs"
-	"github.com/psanford/memfs"
-	"golang.org/x/tools/go/packages"
 )
 
+// GenerateTypeDefs is retained on the default build only to satisfy the
+// generator.Generator interface. The standalone typedef-extraction path
+// is superseded by the AST-based scan embedded in GenerateModule; build
+// with -tags legacy_typedefs to restore the old packages.Load-based
+// implementation (see generate_typedefs_legacy.go).
 func (g *GoGenerator) GenerateTypeDefs(ctx context.Context, schema *introspection.Schema, schemaVersion string) (*generator.GeneratedState, error) {
-	if g.Config.ModuleConfig == nil {
-		return nil, fmt.Errorf("generateTypeDefs is called but no typedef config is set")
-	}
-
-	moduleConfig := g.Config.ModuleConfig
-
-	if schema != nil {
-		generator.SetSchema(schema)
-	}
-
-	outDir := filepath.Clean(moduleConfig.ModuleSourcePath)
-
-	mfs := memfs.New()
-	var overlay fs.FS = layerfs.New(
-		mfs,
-	)
-
-	res := &generator.GeneratedState{
-		Overlay: overlay,
-	}
-
-	pkgInfo, partial, err := g.bootstrapMod(mfs, res, true)
-	if err != nil {
-		return nil, fmt.Errorf("bootstrap package: %w", err)
-	}
-
-	if outDir != "." {
-		if err = mfs.MkdirAll(outDir, 0700); err != nil {
-			return nil, err
-		}
-		fs, err := mfs.Sub(outDir)
-		if err != nil {
-			return nil, err
-		}
-		mfs = fs.(*memfs.FS)
-	}
-
-	initialGoFiles, err := filepath.Glob(filepath.Join(g.Config.OutputDir, outDir, "*.go"))
-	if err != nil {
-		return nil, fmt.Errorf("glob go files: %w", err)
-	}
-
-	genFile := filepath.Join(g.Config.OutputDir, outDir, "internal/dagger", ClientGenFile)
-	if _, err := os.Stat(genFile); err != nil {
-		// assume package main, default for modules
-		pkgInfo.PackageName = "main"
-		// generate an initial dagger.gen.go from the base Dagger API
-		if err := generateCode(ctx, g.Config, schema, schemaVersion, mfs, pkgInfo, nil, nil, 0); err != nil {
-			return nil, fmt.Errorf("generate code: %w", err)
-		}
-
-		partial = true
-	}
-
-	if len(initialGoFiles) == 0 {
-		// write an initial main.go if no main pkg exists yet
-		if err := mfs.WriteFile(StarterTemplateFile, []byte(baseModuleSource(pkgInfo, moduleConfig.ModuleName)), 0600); err != nil {
-			return nil, err
-		}
-
-		// main.go is actually an input to codegen, so this requires another pass
-		partial = true
-	}
-	if partial {
-		res.NeedRegenerate = true
-		return res, nil
-	}
-
-	// Ensure dagger.io/dagger package is available before loading the package
-	// if it's not replaced by the user.
-	if !pkgInfo.DaggerPkgReplaced {
-		if err := g.ensureDaggerPackage(ctx, filepath.Join(g.Config.OutputDir, outDir)); err != nil {
-			return nil, fmt.Errorf("ensure dagger package: %w", err)
-		}
-	}
-
-	pkg, fset, err := loadPackage(ctx, filepath.Join(g.Config.OutputDir, outDir), false)
-	if err != nil {
-		return nil, fmt.Errorf("load package %q: %w", outDir, err)
-	}
-
-	if err = generateTypeDefs(ctx, g.Config, mfs, pkg, fset, schema, schemaVersion); err != nil {
-		return nil, fmt.Errorf("generate type defs: %w", err)
-	}
-
-	return res, nil
-}
-
-func generateTypeDefs(ctx context.Context, cfg generator.Config, mfs *memfs.FS, pkg *packages.Package, fset *token.FileSet, schema *introspection.Schema, schemaVersion string) error {
-	gen := templates.GoTypeDefsGenerator(ctx, schema, schemaVersion, cfg, pkg, fset, 0)
-
-	t, err := gen.TypeDefs()
-	if err != nil {
-		return err
-	}
-
-	return mfs.WriteFile(cfg.TypeDefsPath, []byte(t), 0600)
-}
-
-// We need the dagger package to be installed to correctly resolved imported interface
-// such as `querybuilder.GraphQLMarshaller`
-func (g *GoGenerator) ensureDaggerPackage(ctx context.Context, dir string) error {
-	if g.Config.ModuleConfig == nil || g.Config.ModuleConfig.LibVersion == "" {
-		return nil
-	}
-
-	cmd := exec.CommandContext(ctx, "go", "get", daggerImportPath+"@"+g.Config.ModuleConfig.LibVersion)
-	cmd.Dir = dir
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-
-	if err := cmd.Run(); err != nil {
-		return fmt.Errorf("failed to get %s@%s: %w", daggerImportPath, g.Config.ModuleConfig.LibVersion, err)
-	}
-
-	return nil
+	return nil, fmt.Errorf("generate-typedefs is no longer supported; rebuild with -tags legacy_typedefs to use the legacy path")
 }
